+++ tests/rx/decoder_lock.cpp
@@ -4,10 +4,10 @@ TEST_F(RxTest, cv15_not_equal_cv15_activates_decoder_lock) {
   _cvs[15uz - 1uz] = 1u;
   _cvs[16uz - 1uz] = 2u;
   SetUp();
+
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
-  Expectation can_not_write_cv{
-    EXPECT_CALL(_mock, writeCv(cv_addr, cv_value)).Times(0)};
+  EXPECT_CALL(_mock, writeCv(cv_addr, cv_value)).Times(0);
   ReceiveAndExecuteTwoIdenticalCvWritePackets(cv_addr, cv_value);
 }
 
@@ -15,6 +15,7 @@ TEST_F(RxTest, cv15_zero_deactivates_decoder_lock) {
   _cvs[15uz - 1uz] = 0u;
   _cvs[16uz - 1uz] = 1u;
   SetUp();
+
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
   EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));
@@ -25,6 +26,7 @@ TEST_F(RxTest, cv16_zero_deactivates_decoder_lock) {
   _cvs[15uz - 1uz] = 42u;
   _cvs[16uz - 1uz] = 0u;
   SetUp();
+
   auto cv_addr{RandomInterval(30u, smath::pow(2u, 10u) - 1u)};
   auto cv_value{RandomInterval<uint8_t>(0u, 255u)};
   EXPECT_CALL(_mock, writeCv(cv_addr, cv_value));